requirements.txt
fastapi==0.88.0
SQLAlchemy==1.4.31
uvicorn==0.20.0
docker==6.0.1
python-dotenv==0.21.0
psycopg2-binary==2.9.5
SQLAlchemy-Utils==0.39.0
python-jose==3.3.0
passlib==1.7.4
email-validator==1.1.3
python-multipart==0.0.5
pytest==7.2.0
requests==2.28.1
faker==15.3.4
